feat: integrate AI capabilities into the plugin system
- Introduce a new AI service and repository for managing OpenAI API keys.
- Refactor main.rs to initialize the AI service and update the AI helper to use the new service.
- Modify the plugin host state to include an optional AI helper for enhanced plugin interactions.
- Update the ChatCompletionRequest structure to support larger token limits.
- Remove obsolete AI and logger modules to streamline the plugin architecture.

+++ core/src/dto/openai.rs
@@ -88,7 +88,7 @@ pub struct ChatCompletionRequest {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub model: Option<String>,
     pub messages: Vec<ChatMessage>,
-    pub max_tokens: Option<i32>,
+    pub max_tokens: Option<i64>,
 }
 
 /// Chat completion choice
